Move API wrappers and requests to new Capabilities API bundle and update references.
