+++ extensions/bundles/router.capabilities.api/src/main/java/org/opennaas/extensions/router/capabilities/api/model/chassis/AddInterfacesToLogicalRouterRequest.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package org.opennaas.extensions.router.model.wrappers;
+package org.opennaas.extensions.router.capabilities.api.model.chassis;
 
 /*
  * #%L
@@ -34,7 +34,7 @@ import org.opennaas.extensions.router.model.LogicalPort;
  * @author Jordi
  */
 @XmlRootElement
-public class RemoveInterfacesFromLogicalRouterRequest {
+public class AddInterfacesToLogicalRouterRequest {
 
 	private ComputerSystem		logicalRouter;
 	private List<LogicalPort>	interfaces;
