+++ extensions/bundles/router.capabilities.api/src/main/java/org/opennaas/extensions/router/capabilities/api/model/chassis/RemoveInterfacesFromLogicalRouterRequest.java
@@ -1,4 +1,7 @@
-package org.opennaas.extensions.router.model.wrappers;
+/**
+ * 
+ */
+package org.opennaas.extensions.router.capabilities.api.model.chassis;
 
 /*
  * #%L
@@ -24,18 +27,32 @@ import java.util.List;
 
 import javax.xml.bind.annotation.XmlRootElement;
 
+import org.opennaas.extensions.router.model.ComputerSystem;
 import org.opennaas.extensions.router.model.LogicalPort;
-import org.opennaas.extensions.router.model.OSPFArea;
 
 /**
  * @author Jordi
  */
 @XmlRootElement
-public class AddInterfacesInOSPFAreaRequest {
+public class RemoveInterfacesFromLogicalRouterRequest {
 
+	private ComputerSystem		logicalRouter;
 	private List<LogicalPort>	interfaces;
 
-	private OSPFArea			ospfArea;
+	/**
+	 * @return the logicalRouter
+	 */
+	public ComputerSystem getLogicalRouter() {
+		return logicalRouter;
+	}
+
+	/**
+	 * @param logicalRouter
+	 *            the logicalRouter to set
+	 */
+	public void setLogicalRouter(ComputerSystem logicalRouter) {
+		this.logicalRouter = logicalRouter;
+	}
 
 	/**
 	 * @return the interfaces
@@ -50,21 +67,6 @@ public class AddInterfacesInOSPFAreaRequest {
 	 */
 	public void setInterfaces(List<LogicalPort> interfaces) {
 		this.interfaces = interfaces;
-	}
-
-	/**
-	 * @return the ospfArea
-	 */
-	public OSPFArea getOspfArea() {
-		return ospfArea;
-	}
-
-	/**
-	 * @param ospfArea
-	 *            the ospfArea to set
-	 */
-	public void setOspfArea(OSPFArea ospfArea) {
-		this.ospfArea = ospfArea;
 	}
 
 }
